+++ config/default.rb
@@ -16,7 +16,6 @@ Vagrant.configure("2") do |config|
   config.vm.base_mac = nil
   config.vm.graceful_halt_retry_count = 60
   config.vm.graceful_halt_retry_interval = 1
-  config.vm.guest = :linux
 
   # Share SSH locally by default
   config.vm.network :forwarded_port,