+++ CHANGELOG.md
@@ -30,6 +30,7 @@ BUG FIXES:
   - Finding V1 boxes now works properly again to avoid "box not found"
     errors. [GH-1691]
   - Setting hostname on SLES 11 works again. [GH-1781]
+  - `config.vm.guest` properly forces guests again. [GH-1800]
 
 ## 1.2.2 (April 23, 2013)